pyproject.toml
[project]
name = "noticias"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.13",
]
